[#18] refactor - user context > actions, reducer, index.ts로 분리
