+++ client/src/contexts/user/index.tsx
@@ -1,16 +1,9 @@
 import React, { useEffect, createContext, Dispatch, useReducer, useContext, useState } from 'react';
-import { OK, NOT_FOUND, INTERNAL_SERVER_ERROR } from 'http-status';
-import { UserJoin } from '../../types/data';
 import { UserState } from '../../types/states';
-import * as apis from '../../apis/user';
-import useApiRequest, { REQUEST, SUCCESS, FAILURE } from '../../hooks/useApiRequests';
+import { Action, ACTION_LOGIN_SUCCESS } from './actions';
+import reducer from './reducer';
 
-const ACTION_LOGIN_SUCCESS = 'LOGIN_SUCCESS' as const;
-export const setLoginSuccess = () => ({ type: ACTION_LOGIN_SUCCESS });
-const ACTION_ERROR = 'ERROR' as const;
-type Action = ReturnType<typeof setLoginSuccess>;
-type UserDispatch = Dispatch<Action>;
-
+export type UserDispatch = Dispatch<Action>;
 interface UserContextType {
   state: UserState;
   setAction: UserDispatch;
@@ -23,17 +16,6 @@ const initialState = {
   status: 0, //
 };
 
-function reducer(state: UserState = initialState, action: Action) {
-  switch (action.type) {
-    case ACTION_LOGIN_SUCCESS:
-      return {
-        ...state,
-        isAuthorized: true,
-      };
-    default:
-      return state;
-  }
-}
 const UserContext = createContext<UserContextType | null>(null);
 
 export function UserProvider({ children }: { children: React.ReactNode }) {
